# Improve phonelib performance by reducing object allocations