--- a/lib/phonelib/phone_analyzer.rb
+++ b/lib/phonelib/phone_analyzer.rb
@@ -185,7 +185,7 @@ module Phonelib
       if Phonelib.additional_regexes.is_a?(Hash) && Phonelib.additional_regexes[data[:id]]
         index = Phonelib.additional_regexes[data[:id]].values.flatten.join('|').scan(/\(/).size
       end
-      phone = country_match.to_a[-1 - index]
+      phone = country_match[-1 - index]
       result[:national] = phone
       result[:format] = number_format(phone, data[Core::FORMATS])
       result.merge! all_number_types(phone, data[Core::TYPES], not_valid)