In doing some benchmarking in an application that uses `phonelib`, I saw a few spots that may have some opportunity to improve performance. I wanted to propose changes that were relatively small but could have a larger impact on performance. My approach was to focus more on eliminating allocations where possible by moving otherwise static values into constants or consolidating string allocations. There are a couple of replacements of methods too to prefer less expensive ones (replacing `match` with `match?` and `to_s.length` with `match_length(0)`.

At least locally, the improvements look to be pretty significant:

| benchmark | master | proposed | result |
|-----|-----|-----|----|
| Known Country Memory Allocated | 14.964M | 10.162M | 32% reduction |
| Unknown Country Memory Allocated | 27.413M | 10.542M | 61% reduction |
| Known Country Parsing | 34.552 i/s | 42.581 i/s | 23% increase |
| Unknown Country Parsing |  16.249 i/s | 22.452 i/s | 38% increase |

I have some more significant changes around caching inspired by #47, and will likely open a PR with those. I also wanted to thank you for `phonelib` and its helpful structure/tooling. Having existing benchmarks to test made it really easy to work with!

### Memory usage benchmark
```
# master
$ rspec spec/phonelib_memory_bench.rb
Timing phonelib using Ruby 3.3.6
starting benchmark for 1127 numbers in 245 countries
Calculating -------------------------------------
       known country    14.964M memsize (     2.271M retained)
                       181.611k objects (     6.876k retained)
                        50.000  strings (    50.000  retained)
     unknown country    27.413M memsize (   176.799k retained)
                       363.231k objects (   573.000  retained)
                        50.000  strings (    50.000  retained)

Comparison:
       known country:   14964010 allocated
     unknown country:   27412891 allocated - 1.83x more

# proposed
Timing phonelib using Ruby 3.3.6
starting benchmark for 1127 numbers in 245 countries
Calculating -------------------------------------
       known country    10.162M memsize (     2.272M retained)
                       106.411k objects (     6.876k retained)
                        50.000  strings (    50.000  retained)
     unknown country    10.542M memsize (   176.799k retained)
                       137.597k objects (   573.000  retained)
                        50.000  strings (    50.000  retained)

Comparison:
       known country:   10161695 allocated
     unknown country:   10542221 allocated - 1.04x more
```


### Iterations per second benchmark
```
# master
Timing phonelib using Ruby 3.3.6
starting benchmark for 1127 numbers in 245 countries
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [arm64-darwin23]
Warming up --------------------------------------
       known country     3.000 i/100ms
     unknown country     1.000 i/100ms
Calculating -------------------------------------
       known country     34.552 (± 2.9%) i/s   (28.94 ms/i) -    174.000 in   5.038685s
     unknown country     16.249 (± 6.2%) i/s   (61.54 ms/i) -     82.000 in   5.054568s

Comparison:
       known country:       34.6 i/s
     unknown country:       16.2 i/s - 2.13x  slower

# proposed
Timing phonelib using Ruby 3.3.6
starting benchmark for 1127 numbers in 245 countries
ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [arm64-darwin23]
Warming up --------------------------------------
       known country     4.000 i/100ms
     unknown country     2.000 i/100ms
Calculating -------------------------------------
       known country     42.581 (± 2.3%) i/s   (23.48 ms/i) -    216.000 in   5.073565s
     unknown country     22.452 (± 0.0%) i/s   (44.54 ms/i) -    114.000 in   5.079301s

Comparison:
       known country:       42.6 i/s
     unknown country:       22.5 i/s - 1.90x  slower
```