@daddyz 2025-02-10
@mitchellhenke thanks for PR, however rspec is failing for most of the environments. You can check it [here](https://app.circleci.com/pipelines/github/daddyz/phonelib/142/workflows/d1da4062-6917-4ca2-9065-3343dba004f7). Gem is supporting ruby versions since 2.5 which is one of the issues due to missing methods.
Some of the changes can be accepted but the ones that work in 3+ only can't.