@mitchellhenke 2025-02-10
Ah apologies, I missed that. I'm unfortunately not able to install unsupported versions of Ruby on my machine, so I'll have to rely on the CI for testing those.

I've removed the usage of [MatchData#match_length](https://docs.ruby-lang.org/en/master/MatchData.html#method-i-match_length) which wasn't added until 3.1 and seemed to be the cause of the failures.