@daddyz 2025-02-17
@mitchellhenke not yet, I reconfigured circleci to run tests on pull requests (at least I think so, it was configured before too), now for each your PR it supposed to trigger the pipeline